+++ pkg/controller/servicing/coreos/version.go
@@ -7,9 +7,8 @@ import (
 	"regexp"
 	"time"
 
-	v1 "k8s.io/api/core/v1"
-
 	"github.com/pkg/errors"
+	v1 "k8s.io/api/core/v1"
 
 	"github.com/sapcc/kubernikus/pkg/util/version"
 )
