+++ pkg/controller/servicing/coreos/version_test.go
@@ -1,10 +1,6 @@
 package coreos
 
 import (
-	"bytes"
-	"fmt"
-	"io/ioutil"
-	"net/http"
 	"testing"
 	"time"
 
@@ -51,37 +47,6 @@ const (
 
 	UnexpectedResponse = `404 CoreOS has been bought by IBM`
 )
-
-type RoundTripFunc func(req *http.Request) *http.Response
-
-func (f RoundTripFunc) RoundTrip(req *http.Request) (*http.Response, error) {
-	return f(req), nil
-}
-
-func NewFakeVersion(t *testing.T, version string) *Version {
-	body := fmt.Sprintf("COREOS_VERSION=%s", version)
-	subject := &Version{}
-	subject.Client = NewTestClient(t, "https://stable.release.core-os.net/amd64-usr/current/version.txt", body, nil)
-	return subject
-}
-
-func NewTestClient(t *testing.T, baseURL, body string, count *int) *http.Client {
-	fn := func(req *http.Request) *http.Response {
-		assert.Equal(t, req.URL.String(), baseURL)
-		if count != nil {
-			*count = *count + 1
-		}
-		return &http.Response{
-			StatusCode: 200,
-			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
-			Header:     make(http.Header),
-		}
-	}
-
-	return &http.Client{
-		Transport: RoundTripFunc(fn),
-	}
-}
 
 func TestVersionStable(t *testing.T) {
 	now = func() time.Time { return time.Date(2019, 2, 3, 4, 0, 0, 0, time.UTC) }
